Refactor DefaultObjectGenerator
Introduce PrimitiveValueGenerator class and IntegerGenerator class.
