+++ autoparams/src/main/java/org/javaunit/autoparams/generator/PrimitiveValueGenerator.java
@@ -1,0 +1,11 @@
+package org.javaunit.autoparams.generator;
+
+final class PrimitiveValueGenerator extends CompositeObjectGenerator {
+
+    public PrimitiveValueGenerator() {
+        super(
+            new IntegerGenerator()
+        );
+    }
+
+}
